refactor: isolate torch-free protocol helpers
